Dissociate every filter in Stories
